+++ insanic/permissions.py
@@ -1,5 +1,9 @@
 """
+Copyright © 2011-present, Encode OSS Ltd. All rights reserved.
+
 Provides a set of pluggable permission policies.
+
+Added several permissions for framework usage.
 """
 from insanic.models import _AnonymousUser
 
@@ -11,7 +15,7 @@ class BasePermission(object):
     A base class from which all permission classes should inherit.
     """
 
-    async def has_permission(self, request, view):
+    def has_permission(self, request, view):
         """
         Return `True` if permission is granted, `False` otherwise.
         """
@@ -26,7 +30,7 @@ class AllowAny(BasePermission):
     more explicit.
     """
 
-    async def has_permission(self, request, view):
+    def has_permission(self, request, view):
         return True
 
 
@@ -35,8 +39,8 @@ class IsAuthenticated(BasePermission):
     Allows access only to authenticated users.
     """
 
-    async def has_permission(self, request, view):
-        user = await request.user
+    def has_permission(self, request, view):
+        user = request.user
 
         return not isinstance(user, _AnonymousUser) and user.is_authenticated
 
@@ -46,8 +50,8 @@ class IsAdminUser(BasePermission):
     Allows access only to admin users.
     """
 
-    async def has_permission(self, request, view):
-        user = await request.user
+    def has_permission(self, request, view):
+        user = request.user
         return not isinstance(user, _AnonymousUser) and user.is_staff
 
 
@@ -56,8 +60,8 @@ class IsAuthenticatedOrReadOnly(BasePermission):
     The request is authenticated as a user, or is a read-only request.
     """
 
-    async def has_permission(self, request, view):
-        user = await request.user
+    def has_permission(self, request, view):
+        user = request.user
 
         return request.method in SAFE_METHODS or not isinstance(
             user, _AnonymousUser
@@ -69,9 +73,9 @@ class IsOwnerOrAdmin(BasePermission):
     Custom permission to only allow owners of an object to view or edit it.
     """
 
-    async def has_permission(self, request, view):
+    def has_permission(self, request, view):
 
-        user = await request.user
+        user = request.user
 
         if user.is_staff:
             return True
@@ -87,8 +91,8 @@ class IsAnonymousUser(BasePermission):
     Permission to check this api can only be access by non authenticated user.
     """
 
-    async def has_permission(self, request, view):
-        user = await request.user
+    def has_permission(self, request, view):
+        user = request.user
         return isinstance(user, _AnonymousUser)
 
 
@@ -97,7 +101,7 @@ class IsServiceOnly(BasePermission):
     Permission to check this api can only be access by another service
     """
 
-    async def has_permission(self, request, view):
-        service = await request.service
+    def has_permission(self, request, view):
+        service = request.service
 
         return service.is_authenticated